refactor: move hardcoded resolver out for future integration
