+++ lua/tool-resolver/types.lua
@@ -1,5 +1,11 @@
 ---@class ToolResolver.Config
----@field fallbacks? table<string, string> register fallbacks here
+---@field tools table<string, ToolResolver.Config.Tools> tools with type and fallback
+
+---@class ToolResolver.Config.Tools
+---@field type ToolResolver.ResolverType
+---@field fallback? string fallback binary name, if not specified then use the key.
 
 ---@class ToolResolver.GetBinOpts
 ---@field path? string start search path (default: current buffer)
+
+---@alias ToolResolver.ResolverType "node" -- add more in future
